fungicide/pyproject.toml
[project]
name = "fungicide"
version = "0.1.0"
description = "classification and filtering service for the mycelium web crawler queue"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "joblib>=1.5.1",
    "python-dotenv>=1.1.1",
    "redis[hiredis]>=6.4.0",
    "scikit-learn>=1.7.1",
]
